# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Raffle Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: raffles
    description: Raffle operations
  - name: numbers
    description: Raffle number operations
  - name: buyers
    description: Buyer operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /raffles:
    get:
      operationId: listRaffles
      tags: [raffles]
      summary: List all raffles
      responses:
        "200":
          description: List of raffles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Raffle"
    post:
      operationId: createRaffle
      tags: [raffles]
      summary: Create a new raffle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRaffleInput"
      responses:
        "201":
          description: Created raffle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Raffle"

  /raffles/{id}:
    get:
      operationId: getRaffle
      tags: [raffles]
      summary: Get a raffle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Raffle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaffleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRaffle
      tags: [raffles]
      summary: Update a raffle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRaffleInput"
      responses:
        "200":
          description: Updated raffle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Raffle"
    delete:
      operationId: deleteRaffle
      tags: [raffles]
      summary: Delete a raffle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /raffles/{id}/draw:
    post:
      operationId: drawRaffle
      tags: [raffles]
      summary: Perform the raffle draw
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw result with winners
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawResult"

  /raffles/{id}/numbers:
    get:
      operationId: listRaffleNumbers
      tags: [numbers]
      summary: List all numbers for a raffle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaffleNumber"

  /raffles/{id}/numbers/{number}/assign:
    post:
      operationId: assignNumber
      tags: [numbers]
      summary: Assign a number to a buyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: number
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignNumberInput"
      responses:
        "200":
          description: Updated number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaffleNumber"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /raffles/{id}/numbers/{number}/release:
    post:
      operationId: releaseNumber
      tags: [numbers]
      summary: Release a number back to available
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: number
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaffleNumber"

  /raffles/{id}/buyers:
    get:
      operationId: listBuyers
      tags: [buyers]
      summary: List buyers for a raffle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of buyers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Buyer"

  /raffles/{id}/export/pdf:
    get:
      operationId: exportRafflePdf
      tags: [raffles]
      summary: Export raffle as PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  /raffles/{id}/export/image:
    get:
      operationId: exportRaffleImage
      tags: [raffles]
      summary: Export raffle as PNG image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Raffle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        drawDate:
          type: string
          format: date-time
          nullable: true
        pricePerNumber:
          type: number
        type:
          type: string
          enum: [single_amount, multiple_prizes]
        prizeImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, cancelled]
        prizes:
          type: array
          items:
            type: string
          nullable: true
        singlePrizeAmount:
          type: number
          nullable: true
        totalNumbers:
          type: integer
        soldNumbers:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - pricePerNumber
        - type
        - status
        - totalNumbers
        - soldNumbers
        - createdAt

    RaffleDetail:
      allOf:
        - $ref: "#/components/schemas/Raffle"
        - type: object
          properties:
            numbers:
              type: array
              items:
                $ref: "#/components/schemas/RaffleNumber"
            buyers:
              type: array
              items:
                $ref: "#/components/schemas/Buyer"
            winners:
              type: array
              items:
                $ref: "#/components/schemas/Winner"
              nullable: true

    CreateRaffleInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        drawDate:
          type: string
          format: date-time
          nullable: true
        pricePerNumber:
          type: number
        type:
          type: string
          enum: [single_amount, multiple_prizes]
        prizeImage:
          type: string
          nullable: true
        prizes:
          type: array
          items:
            type: string
          nullable: true
        singlePrizeAmount:
          type: number
          nullable: true
      required:
        - name
        - pricePerNumber
        - type

    UpdateRaffleInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        drawDate:
          type: string
          format: date-time
          nullable: true
        pricePerNumber:
          type: number
        type:
          type: string
          enum: [single_amount, multiple_prizes]
        prizeImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, cancelled]
        prizes:
          type: array
          items:
            type: string
          nullable: true
        singlePrizeAmount:
          type: number
          nullable: true

    RaffleNumber:
      type: object
      properties:
        id:
          type: integer
        raffleId:
          type: integer
        number:
          type: integer
        status:
          type: string
          enum: [available, sold]
        buyerId:
          type: integer
          nullable: true
        buyerName:
          type: string
          nullable: true
      required:
        - id
        - raffleId
        - number
        - status

    Buyer:
      type: object
      properties:
        id:
          type: integer
        raffleId:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        numbers:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - raffleId
        - name
        - numbers
        - createdAt

    AssignNumberInput:
      type: object
      properties:
        buyerName:
          type: string
        buyerPhone:
          type: string
          nullable: true
        buyerEmail:
          type: string
          nullable: true
      required:
        - buyerName

    Winner:
      type: object
      properties:
        number:
          type: integer
        buyerName:
          type: string
        buyerPhone:
          type: string
          nullable: true
        prize:
          type: string
          nullable: true
      required:
        - number
        - buyerName

    DrawResult:
      type: object
      properties:
        raffleId:
          type: integer
        winners:
          type: array
          items:
            $ref: "#/components/schemas/Winner"
      required:
        - raffleId
        - winners