# lib/api-zod/src/generated/types/assignNumberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raffle Management API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignNumberInput {
  buyerName: string;
  buyerPhone?: string | null;
  buyerEmail?: string | null;
}