# lib/api-zod/src/generated/types/buyer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raffle Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Buyer {
  id: number;
  raffleId: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  numbers: number[];
  createdAt: Date;
}