# lib/api-zod/src/generated/types/drawResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raffle Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Winner } from "./winner";

export interface DrawResult {
  raffleId: number;
  winners: Winner[];
}